requirements.txt
Flask~=2.2.2
python-dotenv~=0.21.0
requests~=2.28.1
Werkzeug~=2.2.2
biopython~=1.81
aind_codeocean_api~=0.0.13